feat(ui): Initial organization dashboard
Start adding the organization (security) dashboard by aligning the layout
and component structure first with the product dashboard.

Note that the statistics cards and organization vulnerabilities table
is missing as of yet, due to back-end support not yet ready.
